refactor: add bootstrap as dependency, form field hints, tools folder
