+++ package.json
@@ -8,6 +8,9 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@popperjs/core": "^2.11.6",
+    "bootstrap": "^5.2.3",
+    "bootstrap-icons": "^1.10.3",
     "core-js": "^3.8.3",
     "normalize.css": "^8.0.1",
     "vue": "^3.2.13",
