+++ src/utils/router.js
@@ -1,9 +1,9 @@
 import {createRouter, createWebHashHistory} from 'vue-router';
 
-import MainPage from './components/MainPage.vue';
-import SignUp from './components/SignUp.vue';
-import SignIn from './components/SignIn.vue'
-import EnterScreen from './components/EnterScreen/EnterScreen.vue'
+import MainPage from '../components/MainPage.vue';
+import SignUp from '../components/SignUp.vue';
+import SignIn from '../components/SignIn.vue'
+import EnterScreen from '../components/EnterScreen/EnterScreen.vue'
 
 export default createRouter({
     history: createWebHashHistory(),
