+++ src/utils/tooltip.js
@@ -1,0 +1,10 @@
+/* eslint-disable no-unused-vars */
+import { Tooltip } from 'bootstrap'
+
+export const tooltip = {
+    
+  mounted(el) {
+    const tooltip = new Tooltip(el)
+  }
+
+}
